requirements.txt
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
cloudinary==1.44.1
diff-match-patch==20241021
Django==6.0.2
django-admin-rangefilter==0.13.5
django-cloudinary-storage==0.3.0
django-import-export==4.4.0
django-simple-history==3.11.0
idna==3.11
pillow==12.1.1
qrcode==8.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
tablib==3.9.0
urllib3==2.6.3
whitenoise==6.11.0
